Refactored base classes for data_io, adding checks. Adapted to Maheswaranathan data.
